requirements.txt
GitPython
numpy
pandas
pefile
Pillow
python-magic
scikit-learn
scipy
ssdeep
virustotal-api
yara-python